Fix design issues

+++ src/pages/dashboard/BlogList.js
@@ -2,14 +2,18 @@ import React from 'react'
 import { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { Link } from 'react-router-dom';
+import DefaultLoader from '../../component/DefaultLoader';
 import { deleteBlogData, loadBlogsData } from '../../redux/thunk/fetchBlogData';
 
 function BlogList() {
-  const { blogs } = useSelector(state => state.blogData)
+  const { blogs, loading } = useSelector(state => state.blogData)
   const dispatch = useDispatch();
 
   const handleDelete = (id) => {
-    dispatch(deleteBlogData(id))
+    if(window.confirm('Are you sure??')){
+      dispatch(deleteBlogData(id))
+    }
+   
   }
 
   useEffect(() => {
@@ -24,31 +28,39 @@ function BlogList() {
           <Link to="add-blog" className='dashboard-header-btn btn px-4'>Add Blog</Link>
         </div>
       </div>
-      <table class="table">
-        <thead>
-          <tr>
-            <th scope="col">#</th>
-            <th scope="col">Title</th>
-            <th scope="col">Published At</th>
-            <th className='text-center' scope="col">Action</th>
-          </tr>
-        </thead>
-        <tbody>
-          {
-            blogs && blogs.map((blog, index) => (
-              <tr key={blog.id}>
-              <th scope="row">{index + 1}</th>
-              <td>{blog.title}</td>
-              <td>{blog.date}</td>
-              <td>
-                <Link to={`edit-post/${blog.id}`} className='btn'><i class="fa-solid fa-pen"></i></Link>
-                <button onClick={() => handleDelete(blog.id)} className='btn'><i class="fa-solid fa-trash"></i></button>
-              </td>
-              </tr>
-            ))
-          }
-        </tbody>
-      </table>
+      {
+        loading && <div className="position-relative pb-5 mb-5">
+        <DefaultLoader />
+      </div>
+      }
+
+      <div className='table-responsive'>
+        <table class="table">
+          <thead>
+            <tr>
+              <th scope="col">#</th>
+              <th scope="col">Title</th>
+              <th scope="col">Published At</th>
+              <th className='text-center' scope="col">Action</th>
+            </tr>
+          </thead>
+          <tbody>
+            {
+              blogs && blogs.map((blog, index) => (
+                <tr key={blog.id}>
+                <th scope="row">{index + 1}</th>
+                <td>{blog.title}</td>
+                <td>{blog.date}</td>
+                <td className='text-center'>
+                  <Link to={`edit-post/${blog.id}`} className='btn'><i class="fa-solid fa-pen"></i></Link>
+                  <button onClick={() => handleDelete(blog.id)} className='btn'><i class="fa-solid fa-trash"></i></button>
+                </td>
+                </tr>
+              ))
+            }
+          </tbody>
+        </table>
+      </div>
     </div>
   )
 }
